couple-compass/backend/requirements.txt
# Core FastAPI dependencies
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.8

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background tasks
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# AI and external services
openai==1.3.7
boto3==1.34.0
pinecone-client==3.0.3
websockets==12.0
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.0.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
